daemon: move requirements checks to own file
Have this file owned by @cilium/bpf.

Signed-off by: Ian Vernon <ian@cilium.io>
